requirements.txt
pydantic==2.7.3
aiofiles==23.2.1
uvicorn==0.30.1
fastapi==0.111.0
SQLAlchemy==2.0.30
passlib==1.7.4
PyJWT==2.8.0